Combined the UI folder Refactored GUI code to OnGUI (future practice) Moved debug toggle to Shift+D
