--- Assets/Scripts/UI.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 4ec73fff8aac4491292d42d9383d6fa3
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
